Email parsing passes the class with a C-
In all seriousness, I need to find a way to tighten the feedback loop for
email I/O/.  Might work to either mock an email in postman and hit the server with that,
or figure out a way to route emails to local dev env.  Either way, there's a lot of
wasted time in only triggering via a commit, deploy, and manual email to the
domain.